+++ Pandora/Pattern/Functor.hs
@@ -1,5 +1,6 @@
 module Pandora.Pattern.Functor (module Exports) where
 
+import Pandora.Pattern.Functor.Bivariant as Exports
 import Pandora.Pattern.Functor.Divariant as Exports
 import Pandora.Pattern.Functor.Lowerable as Exports
 import Pandora.Pattern.Functor.Liftable as Exports