+++ CHANGELOG.md
@@ -160,3 +160,4 @@
 * Define `Adaptable` instances for comonad transformer schemes
 * Make `position` and `access` methods of `Store` adaptable
 * Rename `unwrap` method of `Interpreted` type class to `run`
+* Define `Bivariant` functor type class